requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.1
django-allauth==0.35.0
django-cors-headers==2.1.0
django-rest-auth==0.9.3
django-toolbelt==0.0.1
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
